fix: fetch에서 사용하는 interface를 index로 빼냄
여러 fetch에서 공용으로 사용할 수 있기때문에 index 쪽으로 옮기고
사용하는 곳에선 index에있는 interface를 import 하는 방식으로 바꿈
