+++ front/src/fetch/getHistories.ts
@@ -1,16 +1,5 @@
 const baseUrl = "http://localhost:3000/api";
 
-interface HistoryDataType {
-  year: number;
-  month: number;
-  day: number;
-  category: string;
-  paymentMethod: string;
-  income: boolean;
-  amount: number;
-  detail: string;
-}
-
 const getHistories = async (year: number, month: number) => {
   try {
     const response = await fetch(`${baseUrl}/histories/${year}/${month}`, {
@@ -26,4 +15,4 @@ const getHistories = async (year: number, month: number) => {
   }
 };
 
-export { getHistories, HistoryDataType };
+export default getHistories;
